+++ examples/coagents-starter/agent-py/sample_agent/demo.py
@@ -34,11 +34,4 @@ def main():
         host="0.0.0.0",
         port=port,
         reload=True,
-        reload_dirs=(
-            ["."] +
-            (["../../../sdk-python/copilotkit"]
-             if os.path.exists("../../../sdk-python/copilotkit")
-             else []
-             )
-        )
     )